feat(server_monitor): 将 windows 平台特定的代码放在平台独立的文件中, 以使条件编译生效.
创建了一个默认文件, 以使除 windows 以外的平台能够通过编译.
