+++ internal/app/server_monitor/monitor_realtime/server_stat_reaaltime_default.go
@@ -1,0 +1,11 @@
+//go:build !windows
+
+package monitor_realtime
+
+import (
+	"errors"
+)
+
+func getAdaptersInfo() (SystemNetworkAdapterInfoList, error) {
+	return nil, errors.New("not implement!")
+}
